tests/requirements.txt
# Test Dependencies
requests>=2.25.0
PyYAML>=5.4.0
